src/note_updates/requirements.txt
watchdog==6.0.0
PyYAML==6.0
